supression de mon article ajouté

+++ src/app/liste-article/liste-article.component.ts
@@ -5,6 +5,7 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { ArticleService } from '../services/article.service';
 import { ReactiveFormsModule } from '@angular/forms';
+import Swal from 'sweetalert2';
 
 interface Article {
   id: number;
@@ -33,9 +34,22 @@ export class ListeArticleComponent implements OnInit {
       console.log(data);
     });
   }
-   deleteArticle(id: number): void {
-    this.articleService.deleteArticle(id).subscribe(() => {
-      this.articles = this.articles.filter(article => article.id !== id);
+
+  deleteArticle(id:number): void {
+    Swal.fire({
+      title: 'Supprimer l\'article?',
+      text: 'Cette action est définitive.',
+      icon: 'warning',
+      showCancelButton: true,
+      cancelButtonText: 'Annuler',
+      confirmButtonText: 'Supprimer',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.articleService.deleteArticle(id).subscribe(() => {
+          Swal.fire('Success', 'Article supprimé avec succès!', 'success');
+          this.articles = this.articles.filter(article => article.id!== id);
+        });
+      }
     });
   }
 }
